extract sidebar tasks header into own component
